Move zmetadata and zvariable to utils
