+++ xpublish/utils/zarr.py
@@ -2,13 +2,12 @@ import copy
 import logging
 from typing import (
     Any,
-    Dict,
     Optional,
 )
 
+import cachey
 import dask.array
 import numpy as np
-import numpy.typing as npt
 import xarray as xr
 from numcodecs.abc import Codec
 from numcodecs.compat import ensure_ndarray
@@ -37,15 +36,44 @@ ZARR_METADATA_KEY = '.zmetadata'
 logger = logging.getLogger('api')
 
 
+def get_zvariables(
+    dataset: xr.Dataset, cache: cachey.Cache
+):
+    """Returns a dictionary of zarr encoded variables, using the cache when possible."""
+
+    cache_key = dataset.attrs.get(DATASET_ID_ATTR_KEY, '') + '/' + 'zvariables'
+    zvariables = cache.get(cache_key)
+
+    if zvariables is None:
+        zvariables = create_zvariables(dataset)
+
+        # we want to permanently cache this: set high cost value
+        cache.put(cache_key, zvariables, 99999)
+
+    return zvariables
+
+
+def get_zmetadata(
+    dataset: xr.Dataset,
+    cache: cachey.Cache,
+    zvariables: dict,
+):
+    """Returns a consolidated zmetadata dictionary, using the cache when possible."""
+
+    cache_key = dataset.attrs.get(DATASET_ID_ATTR_KEY, '') + '/' + ZARR_METADATA_KEY
+    zmeta = cache.get(cache_key)
+
+    if zmeta is None:
+        zmeta = create_zmetadata(dataset)
+
+        # we want to permanently cache this: set high cost value
+        cache.put(cache_key, zmeta, 99999)
+
+    return zmeta
+
+
 def _extract_dataset_zattrs(dataset: xr.Dataset) -> dict:
-    """Helper function to create zattrs dictionary from Dataset global attrs.
-
-    Args:
-        dataset: The Dataset to extract zattrs from.
-
-    Returns:
-        A dictionary of zattrs.
-    """
+    """helper function to create zattrs dictionary from Dataset global attrs"""
     zattrs = {}
     for k, v in dataset.attrs.items():
         zattrs[k] = encode_zarr_attr_value(v)
@@ -57,14 +85,7 @@ def _extract_dataset_zattrs(dataset: xr.Dataset) -> dict:
 
 
 def _extract_dataarray_zattrs(da: xr.DataArray) -> dict:
-    """Helper function to extract zattrs dictionary from DataArray.
-
-    Args:
-        da: The DataArray to extract zattrs from.
-
-    Returns:
-        A dictionary of zattrs.
-    """
+    """helper function to extract zattrs dictionary from DataArray"""
     zattrs = {}
     for k, v in da.attrs.items():
         zattrs[k] = encode_zarr_attr_value(v)
@@ -81,15 +102,7 @@ def _extract_dataarray_coords(
     da: xr.DataArray,
     zattrs: dict,
 ) -> dict:
-    """Helper function to extract coords from DataArray into a directionary.
-
-    Args:
-        da: The DataArray to extract coords from.
-        zattrs: The zattrs dictionary to add coords to.
-
-    Returns:
-        A dictionary of zattrs with coords added.
-    """
+    '''helper function to extract coords from DataArray into a directionary'''
     if da.coords:
         # Coordinates are only encoded if there are non-dimension coordinates
         nondim_coords = set(da.coords) - set(da.dims)
@@ -104,15 +117,7 @@ def _extract_fill_value(
     da: xr.DataArray,
     dtype: np.dtype,
 ) -> Any:
-    """Helper function to extract fill value from DataArray.
-
-    Args:
-        da: The DataArray to extract fill value from.
-        dtype: The numpy dtype of the DataArray.
-
-    Returns:
-        The fill value of the DataArray.
-    """
+    """helper function to extract fill value from DataArray."""
     fill_value = da.attrs.pop('_FillValue', None)
     return encode_fill_value(fill_value, dtype)
 
@@ -122,16 +127,7 @@ def _extract_zarray(
     encoding: dict,
     dtype: np.dtype,
 ) -> dict:
-    """Helper function to extract zarr array metadata.
-
-    Args:
-        da: The DataArray to extract zarr array metadata from.
-        encoding: The encoding dictionary of the DataArray.
-        dtype: The numpy dtype of the DataArray.
-
-    Returns:
-        A dictionary of zarr array metadata.
-    """
+    """helper function to extract zarr array metadata."""
     meta = {
         'compressor': encoding.get('compressor', da.encoding.get('compressor', default_compressor)),
         'filters': encoding.get('filters', da.encoding.get('filters', None)),
@@ -159,33 +155,20 @@ def _extract_zarray(
     return meta
 
 
-def create_zvariables(dataset: xr.Dataset) -> Dict[str, xr.Variable]:
-    """Helper function to create a dictionary of zarr encoded variables.
-
-    Args:
-        dataset: The Dataset to encode variables from.
-
-    Returns:
-        A dictionary of zarr encoded xarray variables.
-    """
+def create_zvariables(dataset: xr.Dataset) -> dict:
+    """Helper function to create a dictionary of zarr encoded variables."""
     zvariables = {}
 
     for key, da in dataset.variables.items():
-        encoded_da: xr.Variable = encode_zarr_variable(da, name=key)
+        encoded_da = encode_zarr_variable(da, name=key)
         zvariables[key] = encoded_da
 
     return zvariables
 
 
 def create_zmetadata(dataset: xr.Dataset) -> dict:
-    """Helper function to create a consolidated zmetadata dictionary.
+    """Helper function to create a consolidated zmetadata dictionary."""
 
-    Args:
-        dataset: The Dataset to create zmetadata from.
-
-    Returns:
-        A consolidated zmetadata dictionary.
-    """
     zmeta = {
         'zarr_consolidated_format': ZARR_CONSOLIDATED_FORMAT,
         'metadata': {},
@@ -213,14 +196,9 @@ def jsonify_zmetadata(
     dataset: xr.Dataset,
     zmetadata: dict,
 ) -> dict:
-    """Helper function to convert zmetadata dict to a json compatible dict.
+    """Helper function to convert zmetadata dictionary to a json
+    compatible dictionary.
 
-    Args:
-        dataset: The Dataset to convert zmetadata from.
-        zmetadata: The zmetadata dict to convert.
-
-    Returns:
-        A json compatible zmetadata dict.
     """
     zjson = copy.deepcopy(zmetadata)
 
@@ -237,23 +215,11 @@ def jsonify_zmetadata(
 
 
 def encode_chunk(
-    chunk: npt.ArrayLike,
+    chunk: np.typing.ArrayLike,
     filters: Optional[list[Codec]] = None,
     compressor: Optional[Codec] = None,
-) -> npt.ArrayLike:
-    """Helper function largely copied from zarr.Array.
-
-    Args:
-        chunk: The chunk to encode.
-        filters: The filters to apply to the chunk.
-        compressor: The compressor to apply to the chunk.
-
-    Returns:
-        The encoded chunk.
-
-    Raises:
-        RuntimeError: If the chunk's dtype is not an object.
-    """
+) -> np.typing.ArrayLike:
+    """helper function largely copied from zarr.Array"""
     # apply filters
     if filters:
         for f in filters:
@@ -276,18 +242,10 @@ def get_data_chunk(
     da: xr.DataArray,
     chunk_id: str,
     out_shape: tuple,
-) -> npt.ArrayLike:
+) -> np.typing.ArrayLike:
     """Get one chunk of data from this DataArray (da).
 
     If this is an incomplete edge chunk, pad the returned array to match out_shape.
-
-    Args:
-        da: DataArray to get chunk from.
-        chunk_id: Chunk id to get.
-        out_shape: Shape of the output chunk.
-
-    Returns:
-        Chunk of data from this DataArray (da).
     """
     ikeys = tuple(map(int, chunk_id.split('.')))
     if isinstance(da, DaskArrayType):
